steel-fault-detector/backend/requirements.txt
# Steel Plate Fault Detection - Backend Requirements
# Flask and Web Framework
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7

# Machine Learning and Deep Learning
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3
scipy==1.11.2
scikit-learn==1.3.0

# Image Processing
pillow==10.0.0
opencv-python==4.8.0.76

# Data Processing
pandas==2.0.3

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0
